Factored fetchInfo function to retrieve information about the latest update
