+++ updater/updater.go
@@ -15,6 +15,18 @@
 
 package updater
 
+import (
+	"crypto/sha256"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
+	"runtime"
+
+	log "github.com/sirupsen/logrus"
+)
+
 // Start checks if an update has been downloaded and if so returns the path to the
 // binary to be executed to perform the update. If no update has been downloaded
 // it returns an empty string.
@@ -26,4 +38,42 @@ func Start(src string) string {
 // if so downloads it.
 func CheckForUpdates(currentVersion string, updateAPIURL, updateBinURL string, cmdName string) (string, error) {
 	return checkForUpdates(currentVersion, updateAPIURL, updateBinURL, cmdName)
+}
+
+const (
+	plat = runtime.GOOS + "-" + runtime.GOARCH
+)
+
+func fetchInfo(updateAPIURL string, cmdName string) (*availableUpdateInfo, error) {
+	r, err := fetch(updateAPIURL + cmdName + "/" + plat + ".json")
+	if err != nil {
+		return nil, err
+	}
+	defer r.Close()
+
+	var res availableUpdateInfo
+	if err := json.NewDecoder(r).Decode(&res); err != nil {
+		return nil, err
+	}
+	if len(res.Sha256) != sha256.Size {
+		return nil, errors.New("bad cmd hash in info")
+	}
+	return &res, nil
+}
+
+type availableUpdateInfo struct {
+	Version string
+	Sha256  []byte
+}
+
+func fetch(url string) (io.ReadCloser, error) {
+	resp, err := http.Get(url)
+	if err != nil {
+		return nil, err
+	}
+	if resp.StatusCode != 200 {
+		log.Errorf("bad http status from %s: %v", url, resp.Status)
+		return nil, fmt.Errorf("bad http status from %s: %v", url, resp.Status)
+	}
+	return resp.Body, nil
 }
